% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_circlize.R
\name{blast_circlize}
\alias{blast_circlize}
\title{blast double sequence alignment results with circlize}
\usage{
blast_circlize(df, df1)
}
\arguments{
\item{df}{}

\item{df1}{}
}
\value{
a random sample of size
}
\description{
Use the R language package circlize to visualize the blast double sequence alignment results
}
