% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_RIdeogram.R
\name{blast_RIdeogram}
\alias{blast_RIdeogram}
\title{blast double sequence alignment with Rideogram}
\usage{
blast_RIdeogram(df1, df2)
}
\arguments{
\item{df1}{}

\item{df2}{}
}
\value{
residual matrices
}
\description{
Use the R language package Rideogram to display the results of blast double sequence alignment
}
\examples{
\dontrun{
data(data)
attach(data)
re <- Two_Fold_Cross_Validation(a,b)
print(re)
}
}
